refactor(scraping): Extract section capture
Move the raw root-content read to `content.PageContentReader` and the
generic page/overlay capture to `capture.SectionCapture`. The relocation is
mechanical: every wait, timeout, scroll threshold, click ceiling, modal
decision, reference metadata and error boundary is preserved, and the shared
boundaries now go through `ScrapingSession` instead of the facade's imports.
`extract_page` stays on the facade as a delegate.

The white-box tests move with the code into `tests/scraping/test_capture.py`
and `tests/scraping/test_content.py`, retargeted at the canonical owners, and
`tests/test_root_content_dom.py` runs the root-content JavaScript against a
real Chromium DOM, which no mocked `page.evaluate` can cover.

See also: #853
