+++ linkedin_mcp_server/scraping/content.py
@@ -1,0 +1,123 @@
+"""Raw page content reads shared by every scraping workflow."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from linkedin_mcp_server.scraping.session import ScrapingSession
+
+
+class PageContentReader:
+    """Read innerText and raw anchor metadata off the bound page."""
+
+    def __init__(self, session: ScrapingSession):
+        self._session = session
+
+    async def _extract_root_content(
+        self,
+        selectors: list[str],
+    ) -> dict[str, Any]:
+        """Extract innerText and raw anchor metadata from the first matching root."""
+        result = await self._session.page.evaluate(
+            """({ selectors }) => {
+                const normalize = value => (value || '').replace(/\\s+/g, ' ').trim();
+                const containerSelector = 'section, article, li, div';
+                const headingSelector = 'h1, h2, h3';
+                const directHeadingSelector = ':scope > h1, :scope > h2, :scope > h3';
+                const MAX_HEADING_CONTAINERS = 300;
+                const MAX_REFERENCE_ANCHORS = 500;
+
+                const getHeadingText = element => {
+                    if (!element) return '';
+
+                    const heading =
+                        element.matches && element.matches(headingSelector)
+                            ? element
+                            : element.querySelector
+                              ? element.querySelector(directHeadingSelector)
+                              : null;
+
+                    return normalize(heading?.innerText || heading?.textContent);
+                };
+
+                const getPreviousHeading = node => {
+                    let sibling = node?.previousElementSibling || null;
+                    for (let index = 0; sibling && index < 3; index += 1) {
+                        const heading = getHeadingText(sibling);
+                        if (heading) {
+                            return heading;
+                        }
+                        sibling = sibling.previousElementSibling;
+                    }
+                    return '';
+                };
+
+                const root = selectors
+                    .map(selector => document.querySelector(selector))
+                    .find(Boolean);
+                const source = root ? 'root' : 'body';
+                const container = root || document.body;
+                const text = container ? (container.innerText || '').trim() : '';
+                const headingMap = new WeakMap();
+
+                const candidateContainers = [
+                    container,
+                    ...Array.from(container.querySelectorAll(containerSelector)).slice(
+                        0,
+                        MAX_HEADING_CONTAINERS,
+                    ),
+                ];
+                candidateContainers.forEach(node => {
+                    const ownHeading = getHeadingText(node);
+                    const previousHeading = getPreviousHeading(node);
+                    const heading = ownHeading || previousHeading;
+                    if (heading) {
+                        headingMap.set(node, heading);
+                    }
+                });
+
+                const findHeading = element => {
+                    let current = element.closest(containerSelector) || container;
+                    for (let depth = 0; current && depth < 4; depth += 1) {
+                        const heading = headingMap.get(current);
+                        if (heading) {
+                            return heading;
+                        }
+                        if (current === container) {
+                            break;
+                        }
+                        current = current.parentElement?.closest(containerSelector) || null;
+                    }
+                    return '';
+                };
+
+                const references = Array.from(container.querySelectorAll('a[href]'))
+                    .slice(0, MAX_REFERENCE_ANCHORS)
+                    .map(anchor => {
+                        const rawHref = (anchor.getAttribute('href') || '').trim();
+                        if (!rawHref || rawHref === '#') {
+                            return null;
+                        }
+
+                        const href = rawHref.startsWith('#')
+                            ? rawHref
+                            : (anchor.href || rawHref);
+
+                        return {
+                            href,
+                            text: normalize(anchor.innerText || anchor.textContent),
+                            aria_label: normalize(anchor.getAttribute('aria-label')),
+                            title: normalize(anchor.getAttribute('title')),
+                            heading: findHeading(anchor),
+                            in_article: Boolean(anchor.closest('article')),
+                            in_nav: Boolean(anchor.closest('nav')),
+                            in_footer: Boolean(anchor.closest('footer')),
+                        };
+                    })
+                    .filter(Boolean);
+
+                return { source, text, references };
+            }""",
+            {"selectors": selectors},
+        )
+        return result
